+++ mitra_config/src/limits.rs
@@ -6,6 +6,7 @@ use serde::{
 };
 use super::ConfigError;
 
+// Included media types are supported in popular web browsers.
 // Not included
 // https://developer.mozilla.org/en-US/docs/Web/Media/Formats/Image_types
 // - image/tiff only supported by Safari
@@ -14,11 +15,11 @@ const SUPPORTED_MEDIA_TYPES: [&str; 20] = [
     "audio/mpeg",
     "audio/ogg",
     "audio/x-wav",
-    "audio/wav",
-    "audio/vnd.wave",
-    "audio/mp4",
+    "audio/wav", // WAV (used by Chrome)
+    "audio/vnd.wave", // WAV (used by Firefox)
     "audio/aac",
-    "audio/x-m4a",
+    "audio/mp4", // M4A (used by Firefox)
+    "audio/x-m4a", // M4A (used by Chrome)
     "image/apng",
     "image/avif",
     "image/gif",