+++ CHANGELOG.md
@@ -13,6 +13,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 - Added `/api/v1/instance/domain_blocks` API endpoint.
 - Added `filter_blocklist_public` dynamic configuration parameter.
 - Added `database_connection_pool_size` configuration parameter.
+- Allow AAC, M4A and WAV audio uploads ([#180](https://codeberg.org/silverpill/mitra/pulls/180)).
 
 ### Changed